Adjusted Notification UI
