+++ package.json
@@ -111,9 +111,10 @@
     "@babel/preset-flow": "^7.0.0",
     "@babel/preset-react": "^7.0.0",
     "@babel/runtime": "^7.1.5",
-    "@helpscout/artboard": "^0.1.10",
+    "@helpscout/artboard": "0.1.11",
     "@helpscout/helix": "^0.1.0",
     "@storybook/addon-actions": "^4.0.9",
+    "@storybook/addon-knobs": "4.0.9",
     "@storybook/addon-links": "^4.0.9",
     "@storybook/addon-options": "^4.0.8",
     "@storybook/addons": "^4.0.9",
