+++ src/components/Notification/__tests__/Timer.test.js
